requirements.txt
playwright>=1.40.0
aiosqlite>=0.19.0
pyyaml>=6.0
tldextract>=5.0.0
matplotlib>=3.7.0
Pillow>=10.0.0
